wallet: rework xswd server for future relayer compatibility
