+++ xelis_wallet/Cargo.toml
@@ -32,6 +32,7 @@ serde_json = { workspace = true }
 async-trait = { workspace = true }
 chrono = { workspace = true }
 futures = { workspace = true }
+cfg-if = { workspace = true }
 
 # WASM dependencies
 [target.'cfg(target_arch = "wasm32")'.dependencies]
